Fix FS-11873: addStream() adds audio/video tracks in random order
addStream() is a deprecated WebRTC API method, and only works via shimming on some platforms. Internally, it is inconsistent in building an offer SDP in the order which FreeSWITCH expects it, which is audio m-lines first.

This can result in intermittent m-line ordering errors on the client.

Instead addTrack() can be used to force the media tracks to be added to the peer in the order FreeSWITCH expects.